backend/pyproject.toml
[tool.poetry]
name = "flowforce_education"
version = "0.1.0"
description = ""
authors = ["Jan Bijster <janbijster@gmail.com>"]
readme = "README.md"
packages = [{include = "flowforce_education", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
django = "^5.2"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
python-dotenv = "^1.1.0"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.9"
whitenoise = "^6.6.0"
dj-database-url = "^2.3.0"
uvicorn = "^0.34.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
